refactor: Update frontend with endpoint name change
